{Update} Shield Fonctionnement
Rajout d'une barre de vie au shield automatiquement mis à jour suivant sa vie actuel.

Mise au point de Coroutine pour gérer la régénération du shield ainsi que de sa réparation.

Changement de la fonction PerformUpgrade en protected.

+++ Assets/Scripts/Fights/CanTakeDamage.cs
@@ -1,3 +1,3 @@
 public interface ICanTakeDamage {
-    public void TakeDamage(int damage); //DEBUG!!! l'appliquer aux ennemis
+    public void TakeDamage(float damage); //DEBUG!!! l'appliquer aux ennemis
 }
